AuthRedis연동 세션 관리로 변경 이후 World 에서도 Redis 연결 고려 AuthServer 를 병렬 사용하기위한 조치 중복 로그인 관리용으로도 필요
TODO : Login기록 및 데이터 log 및 DB화 진행 (LogServer?)

+++ AuthServer/Session/AuthDBSession.cs
@@ -34,7 +34,7 @@ namespace AuthServer.Session
             Send(new ArraySegment<byte>(sendBuffer));
         }
 
-        public void SendLoginRequest(string accountId, int sessionId)
+        public void SendLoginRequest(string accountId, long sessionId)
         {
             var packet = new AdGetAccountVerifyInfo
             {
