+++ Commons/Server.Core/SecurePacketServerSession.cs
@@ -10,7 +10,7 @@ namespace Server.Core
 {
     public abstract class SecurePacketServerSession : ISession
     {
-        public int SessionId { get; set; }
+        public long SessionId { get; set; }
         public Action<Exception>? ErrorHandler { get; set; }
 
         private readonly X509Certificate2 serverCertificate;
